+++ src/adapter/graphql/src/scalars/dataset/adapters/extra_data.rs
@@ -9,6 +9,8 @@
 
 use std::sync::LazyLock;
 
+use kamu_datasets::ExtraDataFields;
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 static EXTRA_DATA_JSONSCHEMA_VALIDATOR: LazyLock<jsonschema::Validator> = LazyLock::new(|| {
@@ -82,3 +84,9 @@ fn invalid_input_error(
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+impl From<ExtraData> for ExtraDataFields {
+    fn from(value: ExtraData) -> Self {
+        Self::new(value.into_inner())
+    }
+}